update to output prometheus metrics

+++ pkg/types/types.go
@@ -2,5 +2,5 @@ package types
 
 type DsTypeCount struct {
 	DsType  string
-	DsCount int64
+	DsCount float64
 }
